Separate comfyui operators
